requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
apscheduler>=3.10.0
pyyaml>=6.0
httpx>=0.24.0
openai>=1.0.0
